% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_specie.R
\name{add_specie}
\alias{add_specie}
\title{add_specie}
\usage{
add_specie(species_to_add, dbcon, commit = FALSE)
}
\arguments{
\item{species_to_add}{names of species to add to specie table (character vector)}

\item{dbcon}{Database connection, as created by [RPostgres::dbConnect]}

\item{commit}{value indicating whether to commit database transaction
(boolean: TRUE or FALSE; default is FALSE)}
}
\value{
exits silently on success
}
\description{
Add a specie to BETYdb species table
}
\examples{
\dontrun{
add_specie(species_to_add = c('Sorghum bicolor'), dbcon = dbcon,
commit = FALSE)
}
}
