% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-metadata.R
\name{add_species}
\alias{add_species}
\title{Create new species entry to SPI-Birds}
\usage{
add_species(taxa)
}
\arguments{
\item{taxa}{List of taxa, eq of list output of \link{convert_to_eml}.}
}
\value{
A data frame with a number of rows equal to the number of species to be added, and variables equal to the structure of \link{species_codes}.
}
\description{
Create new species entry for the SPI-Birds species code table (\link{species_codes}).
}
\details{
Taxon ids are generated through \link{get_taxon_ids} as part of \link{convert_to_eml}. This function generates SPI-Birds internal ids. \code{speciesCode} is a running number; \code{specisID} is a 6-letter code created from the first three letters of the genus and thre first three letters of the specific epithet. If this code already exists, the user is prompted to provide a unique code. Vernacular names are taken from Wikidata and GBIF; if multiple English vernacular names are in use, the user is prompted to pick one. Tip: use the Common Names section of a taxon on EOL to find the 'preferred name'.

Only used within \link{add_metadata}.
}
