% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-study-site.R
\name{add_study_site_metadata}
\alias{add_study_site_metadata}
\title{Add study & study site of new metadata entry to SPI-Birds tables}
\usage{
add_study_site_metadata(meta, primaryData = FALSE, formattedData = NA)
}
\arguments{
\item{meta}{Output from \code{convert_to_eml()}}

\item{primaryData}{Logical indicating whether primary data were submitted with the metadata. Default: FALSE.}

\item{formattedData}{Logical indicating whether submitted primary data were standardised. Can only be set to TRUE/FALSE when \code{primaryData} is TRUE. Default: NA.}
}
\description{
Add study & study site of new metadata entry to SPI-Birds tables
}
