\name{AIRPORT.FILE}
\alias{AIRPORT.FILE}
\docType{data}
\title{A file of airport locations
 
}
\description{This is the filename for the airport locations file
 
}
\usage{ AIRPORT.FILE }
\format{
  The format is:
 chr "airports.csv"
}
\details{ The file is compiled from a variety of sources and updated
  frequently. It is used here primarily to get the locations of
  airports.
 
}
\source{
\url{http://www.ourairports.com/}
}
\references{ secondary sources are listed here
  \url{http://www.ourairports.com/about.html#credits}
  They include the FAA and several dedicated individuals.
 
}
\examples{
print(AIRPORT.FILE)
 
}
 
\keyword{filenames}
