\name{AIRPORTS_DIR}
\alias{AIRPORTS_DIR}
\docType{data}
\title{ Directory name for Aiport data
 
}
\description{Directory where aiport files are kept
 
}
\usage{AIRPORTS_DIR }
\format{
  The format is:
 chr "Airports"
}
 
\examples{
print(AIRPORTS_DIR)
 
}


\keyword{directories}
