% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-metadata.R
\name{assign_custodianID}
\alias{assign_custodianID}
\title{Manually assign an ID for the custodian}
\usage{
assign_custodianID(meta)
}
\arguments{
\item{meta}{List output of \link{convert_to_eml}.}
}
\description{
Pick an ID for the custodian of a new metadata entry. custodianIDs are often based on abbreviations of the organisation.
For example, 'LDZ' is the custodianID for 'University of Lodz' and 'TRN' is the custodianID for 'University of Turin'.
The function checks whether the provided ID is already assigned to another study, and will ask you whether you wish to assign the
same custodian to the new entry, or otherwise, pick a new unique ID.
}
\details{
Only used within \link{add_metadata}.
}
