% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tags.R
\name{assign_tags}
\alias{assign_tags}
\title{Assigns tags to deployed content in RStudio Connect.}
\usage{
assign_tags(valid_rmd_tags, rsc_tags, content_guid)
}
\arguments{
\item{valid_rmd_tags}{The output from \code{validate_rmd_tags}; a list with the tags available both in the Rmd and RSC.}

\item{rsc_tags}{The output from \code{get_all_rsc_tags}; an unordered list with the tags in RSC.}

\item{content_guid}{The unique identifier of the desired content item.}
}
\description{
Assigns tags to deployed content in RStudio Connect.
}
\keyword{internal}
