% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate_metadata.R
\name{associate_metadata}
\alias{associate_metadata}
\alias{associate_experiments_sites}
\alias{associate_experiments_treatments}
\alias{associate_sites_cultivars}
\alias{associate_citations_sites}
\title{Associate metadata added to BETYdb}
\usage{
associate_experiments_sites(dbcon, commit, sites)

associate_experiments_treatments(dbcon, commit, treatments)

associate_sites_cultivars(dbcon, commit, sites)

associate_citations_sites(dbcon, commit, citations, sites)
}
\arguments{
\item{dbcon}{Database connection, as created by [RPostgres::dbConnect]}

\item{commit}{Value indicating whether to commit database transactions (boolean, TRUE/FALSE)}

\item{sites}{Sites google sheet object}

\item{treatments}{Treatments google sheet object}

\item{citations}{Citations google sheet object}
}
\description{
After new metadata has been added to BETYdb, associations must be made between experiments
and sites, experiments and treatments, sites and cultivars, and citations and sites.
}
\examples{
\dontrun{

library(RPostgres)
library(googlesheets)
library(DBI)

# connect to test database
dbcon <- dbConnect(RPostgres::Postgres(),
host = 'localhost',
user = 'bety',
port = 5433)

# url to google sheet
url <- 'https://docs.google.com/spreadsheets/d/
1BPkZPfynZI9DB9bRJfa335uC_uNjST5jHVf-XUgSVvM/edit#gid=347205052'

# get key from URL
key <- extract_key_from_url(url)

# register sheet
gs_obj <- key \%>\% gs_key(lookup = FALSE)

# sites sheet
sites <- gs_obj \%>\% gs_read(ws = 'sites')

# associate experiments with sites in BETYdb, commit database transaction
associate_experiments_sites(sites, dbcon, commit = TRUE)

}


}
