% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate_vals.R
\name{associate_vals}
\alias{associate_vals}
\title{associate_vals}
\usage{
associate_vals(tbl_name, col_1, col_2, val_1, val_2, dbcon,
  commit = FALSE)
}
\arguments{
\item{tbl_name}{Name of BETYdb table (character, length 1)}

\item{col_1}{Name of first id column of BETYdb table (character length 1)}

\item{col_2}{Name of second id column of BETYdb table (character length 1)}

\item{val_1}{First id value corresponding to column 1 (integer length 1)}

\item{val_2}{Second id value corresponding to column 2 (integer length 1)}

\item{dbcon}{Database connection, as created by [RPostgres::dbConnect]}

\item{commit}{Value indicating whether to commit database transaction;
(boolean: TRUE or FALSE, default is FALSE)}
}
\value{
Exits on success
}
\description{
Associate two values in BETYdb
}
\examples{
\dontrun{
associate_vals(tbl_name = 'experiment_sites', col_1 = 'experiment_id',
col_2 = 'site_id', val_1 = 900000004, val_2 = 900000178, dbcon = dbcon,
commit = FALSE)
}
}
