\name{BLUEWATER.URL}
\alias{BLUEWATER.URL}
\docType{data}
\title{Url to blue water irrigation data
 
}
\description{"Highest blue water consumption rates occur in warm 
   regions where green water (stemming directly from precipitation)
   is not sufficient to sustain optimal crop growth and where enough
   blue water is available to fulfill the resulting irrigation
   water requirements of the crops. Note that the values refer to 
   cropland areas rather than to entire 0.5 deg grid cells, 
   i.e. the influence of the fractional coverage of a grid cell 
   with cropland is eliminated."
 
}
\usage{ BLUEWATER.URL }
\format{
  The format is:
 chr "http://atlas.gwsp.org/atlas/data/bluewater1_0.zip"
}
 
\source{\url{http://atlas.gwsp.org/index.php}
 Please cite this map as: "GWSP Digital Water Atlas (2008). 
  Map 64: Blue Water Consumption on Cropland (V1.0). 
  Available online at http://atlas.gwsp.org."
 
}
\references{Dr. Dieter Gerten, Potsdam-Institute for Climate Impact Research,
 Earth System Analysis Research Domain,
  P.O. Box 60 12 03, 14412 Potsdam, Germany; 
  E-mail: gerten@pik-potsdam.de
 
}
 
 
\examples{
print(BLUEWATER.URL)
 
}
 
\keyword{urls}
