\name{BLUEWTR_DIR}
\alias{BLUEWTR_DIR}
\docType{data}
\title{ bluewater datafile directory
 
}
\description{ The Bluewater directory contains the
  datafiles for bluewater irrigation
 
}
\usage{ BLUEWTR_DIR }
\format{
  The format is:
 chr "Bluewater"
}
 
 
\examples{
print(BLUEWTR_DIR)
 
}
 
\keyword{directories}
