% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_content.R
\name{bundle_content}
\alias{bundle_content}
\title{Creates the bundle to be pushed to RStudio Connect.}
\usage{
bundle_content(rmd_file_path, content_guid, files_to_exclude = NULL)
}
\arguments{
\item{rmd_file_path}{File path to the RMarkdown document}

\item{content_guid}{The unique identifier of the desired content item}

\item{files_to_exclude}{Files you don't want inside of the bundle}
}
\value{
bundle_id
}
\description{
Creates the bundle to be pushed to RStudio Connect.
}
\keyword{internal}
