\name{COVER.TYPE}
\alias{COVER.TYPE}
\docType{data}
\title{ text translations of land cover types
 
}
\description{ This collection of text descriptions of land cover types
  is based on LCCS, the land cover classification system. It is used
  to translate the numeric codes in the dataset
 
}
\usage{ COVER.TYPE }
\format{
  The format is:
 chr [1:21] "Broadleaf Evergreen Forest" "Broadleaf Deciduous Forest" ...
}
\details{The Global Land Cover by National Mapping Organizations (GLCNMO)
   dataset is coded using LCCS. This differs slightly from the Modis
   landcover product
}
 
\references{
  url{http://www.fao.org/docrep/003/X0596E/X0596e00.HTM}
}
\examples{
print(COVER.TYPE)
 
}

\keyword{constants}
