% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-map.R
\name{create_map}
\alias{create_map}
\title{Create leaflet map of study site}
\usage{
create_map(
  site = NULL,
  zoom = 12,
  tiles = "OpenStreetMap",
  file_format = c("png", "html")
)
}
\arguments{
\item{site}{Character indicating the siteID of the site. If NULL, a map will be generated for all sites.}

\item{zoom}{Integer indicating the zoom level. See \link[leaflet]{setView}.}

\item{tiles}{Character indicating the source for the base map. One of \link[leaflet]{providers}.}

\item{file_format}{Character indicating the format to save in. Either 'png' or 'html'.}
}
\description{
Create a static image of an interactive leaflet map for the SPI-Birds website.
}
