% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-party.R
\name{create_party}
\alias{create_party}
\title{Create an additional party}
\usage{
create_party()
}
\description{
The Responsible Party (creator), Metadata Provider, and Contact in Jotform are limited to a single entry. In some instances, data contributors wish to add more than one party to these entities. This function will help with the creation of that party.
}
