% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rsc_content.R
\name{create_rsc_content}
\alias{create_rsc_content}
\title{Function to create the content in RStudio Connect}
\usage{
create_rsc_content(
  name,
  description,
  title,
  access_type = "acl",
  connection_timeout = NULL,
  idle_timeout = NULL,
  init_timeout = NULL,
  load_factor = NULL,
  max_conns_per_process = NULL,
  max_processes = NULL,
  min_processes = NULL,
  read_timeout = NULL
)
}
\arguments{
\item{name}{Name of the application to be used as an identifier}

\item{description}{A rich description of the content}

\item{title}{The title of the content}

\item{access_type}{Access controls for the content. Allowed options are: "acl" (default), "logged_in", "all"}

\item{connection_timeout}{Maximum number of seconds allowed without data sent or received across a client connection}

\item{idle_timeout}{The maximum number of seconds a worker process for an interactive application to remain alive after it goes idle}

\item{init_timeout}{The maximum number of seconds allowed for an interactive application to start}

\item{load_factor}{Controls how aggressively new processes are spawned}

\item{max_conns_per_process}{Specifies the maximum number of client connections allowed to an individual process.}

\item{max_processes}{Specifies the total number of concurrent processes allowed for a single interactive application}

\item{min_processes}{Specifies the minimum number of concurrent processes allowed for a single interactive application}

\item{read_timeout}{Maximum number of seconds allowed without data received from a client connection}
}
\value{
content_guid
}
\description{
Function to create the content in RStudio Connect
}
\details{
You can take a look at the options here: https://docs.rstudio.com/connect/1.8.6/api/#post-/v1/content
}
\keyword{internal}
