\name{CULTIVATED.FILE}
\alias{CULTIVATED.FILE}
\docType{data}
\title{ File name of land use data
 
}
\description{The file name of the land use data that represents
  the cultivated land.
 
}
\usage{ CULTIVATED.FILE }
\format{
  The format is:
 chr "CULT_2000.asc"
}
 
\source{
 url{http://www.iiasa.ac.at/Research/LUC/External-World-soil-database/
 HTML/LandUseSharesDesc.html?sb=8}
}
\references{Fischer, G., F. Nachtergaele, S. Prieler, H.T. van Velthuizen,
 L. Verelst, D. Wiberg, 2008. Global Agro-ecological Zones Assessment for 
 Agriculture (GAEZ 2008). IIASA, Laxenburg, Austria and FAO, Rome, Italy.
  
}
\examples{
print(CULTIVATED.FILE)
 
}
\keyword{filenames}
