% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_bundle.R
\name{deploy_bundle}
\alias{deploy_bundle}
\title{Function to deploy the bundle in RStudio Connect}
\usage{
deploy_bundle(bundle_id, content_guid)
}
\arguments{
\item{bundle_id}{Bundle identifier within RSC}

\item{content_guid}{The unique identifier of the desired content item}
}
\value{
httr response object
}
\description{
Function to deploy the bundle in RStudio Connect
}
\keyword{internal}
