% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{euring_codes}
\alias{euring_codes}
\title{EURING bird species codes}
\format{
A data frame
\describe{
\item{Status}{Character indicating the status of the code. a = aggregate of (sub)species; f = feral species; h = hybrid species; o = obsolete code; sp = species; ssp = subspecies}
\item{EURING_Code}{5-digit character indicating the EURING code for a (sub)species}
\item{Current_Name}{Current scientific name for (sub)species}
\item{Date_Updated}{Date since the code was last updated}
\item{Notes}{Any additional information on (sub)species}
\item{Old_Name}{Old scientific name for (sub)species}
\item{IOC14_1}{IOC 14.1 code for (sub)species}
}
}
\usage{
euring_codes
}
\description{
Species name, EURING code, date last updated.
}
\keyword{datasets}
