\name{FOREST.FILE}
\alias{FOREST.FILE}
\docType{data}
\title{ File name for land use data 
 
}
\description{ The name of the land use file that contains data
  about the percentage of forest land
 
}
\usage{ FOREST.FILE }
\format{
  The format is:
 chr "FOR_2000.asc"
}
 
\source{
  url{http://www.iiasa.ac.at/Research/LUC/External-World-soil-database/
  HTML/LandUseShares.html?sb=9}
}
\references{
   Fischer, G., F. Nachtergaele, S. Prieler, H.T. van Velthuizen, 
   L. Verelst, D. Wiberg, 2008. Global Agro-ecological Zones Assessment 
   for Agriculture (GAEZ 2008). IIASA, Laxenburg, Austria and FAO, Rome, Italy.
}
\examples{
print(FOREST.FILE)
 
}
 
\keyword{filenames}
