% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{get_fields}
\usage{
get_fields(tbl_fields, row)
}
\arguments{
\item{tbl_fields}{Names of fields of specific betydb table that are also field names of row;
used to subset row (character vector)}

\item{row}{Single row of a google sheet (tibble with dim 1 x n)}
}
\value{
Names of fields of row that have a user inputted value
and is a field name in specific betydb table (character vector)
}
\description{
Get names of fields for a row that have a value
and is also a field name in betydb table
}
\examples{
\dontrun{
get_fields(tbl_fields = c('name', 'start_date', 'end_date', 'user_id'),
row = tibble(name = 'MAC Season 8: Border', start_date = '2019-01-01', end_date = '2019-03-01',
description = NA, design = NA))
}
}
