% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_cultivars.R
\name{get_new_cultivars}
\alias{get_new_cultivars}
\title{Get new cultivars}
\usage{
get_new_cultivars(dbcon, cultivars)
}
\arguments{
\item{dbcon}{Database connection, as created by [RPostgres::dbConnect]}

\item{cultivars}{Cultivars google sheet object}
}
\value{
Combinations of cultivar and specie id to be added to BETYdb cultivars table (tibble)
}
\description{
Only combinations of cultivar and specie id that have not yet been added to the BETYdb
cultivars table should be added when running add_cultivars(). get_new_cultivars() will
return subset of cultivars and specie ids from google sheet that are not present in
BETYdb and that should be added.
}
\details{
The output of this function should be passed into add_cultivars().
}
\examples{
\dontrun{

library(RPostgres)
library(googlesheets)
library(DBI)

# connect to test database
dbcon <- dbConnect(RPostgres::Postgres(),
host = 'localhost',
user = 'bety',
port = 5433)

# url to google sheet
url <- 'https://docs.google.com/spreadsheets/d/
1BPkZPfynZI9DB9bRJfa335uC_uNjST5jHVf-XUgSVvM/edit#gid=347205052'

# get key from URL
key <- extract_key_from_url(url)

# register sheet
gs_obj <- key \%>\% gs_key(lookup = FALSE)

# cultivars sheet
cultivars <- gs_obj \%>\% gs_read(ws = 'cultivars')

# get new cultivars to add to BETYdb
get_new_cultivars(dbcon, cultivars)

}
}
