% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{get_params}
\usage{
get_params(row, fields)
}
\arguments{
\item{row}{Single row from google sheet (tibble with dim 1 x n)}

\item{fields}{Names of fields for returned parameters;
to be used to subset row (character vector)}
}
\value{
Parameters (unnamed list)
}
\description{
Get parameters for a prepared statement
}
\examples{
\dontrun{
get_params(row = tibble(name = 'MAC Season 8: Border', start_date = '2019-01-01',
end_date = '2019-03-31', description = NA, design = NA),
fields = c('name', 'start_date', 'end_date'))
}
}
