% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_tags.R
\name{get_rmd_tags}
\alias{get_rmd_tags}
\title{Function to extract all tags from the RMarkdown document}
\usage{
get_rmd_tags(rmd_file_path)
}
\arguments{
\item{rmd_file_path}{File path to the RMarkdown document}
}
\value{
The tags present in the RMarkdown document.
}
\description{
Function to extract all tags from the RMarkdown document
}
\keyword{internal}
