% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statement.R
\name{get_statement}
\alias{get_statement}
\title{get_statement}
\usage{
get_statement(dbcon, tbl_name, fields)
}
\arguments{
\item{dbcon}{Database connection, as created by [RPostgres::dbConnect]}

\item{tbl_name}{Name of BETYdb table (character, length 1)}

\item{fields}{Names of fields to include in statement (character vector)}
}
\value{
Parameterized sql statement to be passed into `prepared_statement` (character, length 1)
}
\description{
Get parameterized statement to insert data into specified BETYdb table
}
\examples{
\dontrun{
get_statement(dbcon = dbcon, tbl_name = 'experiments',
fields = c('name', 'start_date', 'end_date', 'user_id'))
}
}
