% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-metadata.R
\name{get_vernacular_name}
\alias{get_vernacular_name}
\title{Get English vernacular name of a species}
\usage{
get_vernacular_name(species)
}
\arguments{
\item{species}{The scientific name of the species. Format: '\if{html}{\out{<genus>}} \if{html}{\out{<specific epithet>}}'. For example, 'Parus major'.}
}
\description{
Get English vernacular name of a species from Wikidata and GBIF Backbone Taxonomy. If multiple vernacular names are used, the user is prompted to select one. If neither Wikidata nor GBIF provides a common name, user is prompted to select EOL common name.
}
\details{
Only used within \link{add_species}, and with that, \link{add_metadata}.
}
