\name{GMT.FILE}
\alias{GMT.FILE}
\docType{data}
\title{ Filename for distance from coast data
 
}
\description{ Filename for distance from coast data
 }
\usage{ GMT.FILE }
\format{
  The format is:
 chr "GMT_intermediate_coast_distance_01d.tif"
}
 
\source{
 url{http://oceancolor.gsfc.nasa.gov/DOCS/DistFromCoast/}
}
\references{
 
  Produced by 
  Rick Stumpf (NOAA, National Ocean Service) 
  richard.stumpf at noaa.gov
  and Norman Kuring at the NASA Ocean Biology Processing Group,
  Norman.A.Kuring at nasa.gov
  June 2009

  based on concept of Eric Leuliette of NOAA Satellite Altimetry Lab. 
}
\examples{
print(GMT.FILE)
 
}
 
\keyword{filenames}
