\name{GRASSLAND.FILE}
\alias{GRASSLAND.FILE}
\docType{data}
\title{ Filename for land use data
 
}
\description{ Filename for land use data. Percent of land covered by grass
 
}
\usage{ GRASSLAND.FILE }
\format{
  The format is:
 chr "GRS_2000.asc"
}
 
\examples{
print(GRASSLAND.FILE)
 
}
 
\keyword{filenames}
