\name{GRMPURB_DIR}
\alias{GRMPURB_DIR}
\docType{data}
\title{ Directory for GRUMP urban extent data
 
}
\description{ Directory for GRUMP urban extent
 
}
\usage{ GRMPURB_DIR }
\format{
  The format is:
 chr "GrumpUrban"
}
 
\examples{
print(GRMPURB_DIR)
 
}
 
\keyword{directories}
