% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{habitat_codes}
\alias{habitat_codes}
\title{EUNIS habitat codes}
\format{
A data frame
\describe{
\item{habitatID}{Identifier for the habitat as provided by EUNIS.}
\item{habitatLevel}{Hierarchical level for the habitat as provided by EUNIS. Options: 1-8}
\item{habitatType}{Name or short description for the habitat as proved by EUNIS.}
\item{habitatDetails}{Detailed description or definition for the habitat as provided by EUNIS.}
}
}
\usage{
habitat_codes
}
\description{
Habitat classification (version 2012) according to the European Nature Information System (EUNIS). EUNIS habitat classification is hierarchical (up to eight levels).
}
\keyword{datasets}
