\name{HYDE.URBAN.RASTER}
\alias{HYDE.URBAN.RASTER}
\docType{data}
\title{ File name for raster of hyde urban population data
 
}
\description{File name for raster of Hyde3.1 urban population data
 
}
\usage{ HYDE.URBAN.RASTER }
\format{
  The format is:
 chr "HydeUrban.grd"
}
\details{details are covered in the primary literature and summarized
  at this url \url{http://themasites.pbl.nl/en/themasites/hyde/
  basicdrivingfactors/population/index.html}
  The data is presented by decade in 5 minute bins.
  
 }
\source{\url{http://themasites.pbl.nl/en/themasites/hyde/index.html}
  
}
\references{Klein Goldewijk, K. , A. Beusen, M. de Vos and G. van Drecht (2011). 
  The HYDE 3.1 spatially explicit database of human induced land use 
  change over the past 12,000 years, Global Ecology and 
  Biogeography20(1): 73-86. DOI: 10.1111/j.1466-8238.2010.00587.x.
  
}
\examples{
print(HYDE.URBAN.RASTER)
 
}
 
\keyword{rasters}
