\name{IRRIGATED.FILE}
\alias{IRRIGATED.FILE}
\docType{data}
\title{ Filename for irrigation data
 
}
\description{filename for irrigation data
 
}
\usage{ IRRIGATED.FILE }
\format{
  The format is:
 chr "CULTIR_2000.asc"
}
 
\examples{
print(IRRIGATED.FILE)
 
}
 
\keyword{filenames}
