\name{LANDUSE.RASTER}
\alias{LANDUSE.RASTER}
\docType{data}
\title{ filename for a raster of all the land use source files
 
}
\description{The various land use files, irrigation, cultivation, forest
  grass, sparse vegatation, and urban are all collated into one raster "brick"
  
}
\usage{ LANDUSE.RASTER }
\format{
  The format is:
 chr "landuse.grd"
}
 
 
\examples{
print(LANDUSE.RASTER)
 
}
 
\keyword{rasters}