\name{LIGHTS.FILE}
\alias{LIGHTS.FILE}
\docType{data}
\title{ filename of the nightlights file
 
}
\description{filename of the nightlights file
 
}
\usage{ LIGHTS.FILE }
\format{
  The format is:
 chr "F16_20051128_20061224.cloud2.light1.marginal0.fg_15_35_55.
      full_swath.line_screened.shiftx2y2.op_blended.avg_vis.tif"
}
 
\examples{
print(LIGHTS.FILE)
 
}
 
\keyword{filenames}

