% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_publish_rmd.R
\name{meta_publish_rmd}
\alias{meta_publish_rmd}
\title{Function to publish an Rmd document with tags to RStudio Connect}
\usage{
meta_publish_rmd(rmd_file_path, name, description, title, ...)
}
\arguments{
\item{rmd_file_path}{File path to the RMarkdown document}

\item{name}{Name of the application to be used as an identifier}

\item{description}{A rich description of the content}

\item{title}{The title of the content}

\item{...}{All other parameters option to creating content in RStudio Connect. You can take a look at the options here: https://docs.rstudio.com/connect/1.8.6/api/#post-/v1/content}
}
\description{
Function to publish an Rmd document with tags to RStudio Connect
}
