% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataPerson.R
\name{MetadataPerson}
\alias{MetadataPerson}
\title{Shiny interface for managing metadata for a person}
\description{
A modular shiny interface for managing metdata for a person
}
\section{Super class}{
\code{\link[metadata:OrderedListElement]{metadata::OrderedListElement}} -> \code{MetadataPerson}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetadataPerson$new()}}
\item \href{#method-copyData}{\code{MetadataPerson$copyData()}}
\item \href{#method-createShinyUI}{\code{MetadataPerson$createShinyUI()}}
\item \href{#method-updateShinyUI}{\code{MetadataPerson$updateShinyUI()}}
\item \href{#method-createShinyServer}{\code{MetadataPerson$createShinyServer()}}
\item \href{#method-clone}{\code{MetadataPerson$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$new(
  name,
  templatePerson = NULL,
  role = NULL,
  funding = FALSE,
  copy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copyData"></a>}}
\if{latex}{\out{\hypertarget{method-copyData}{}}}
\subsection{Method \code{copyData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$copyData(metadataPerson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-createShinyUI}{}}}
\subsection{Method \code{createShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$createShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-updateShinyUI}{}}}
\subsection{Method \code{updateShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$updateShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createShinyServer"></a>}}
\if{latex}{\out{\hypertarget{method-createShinyServer}{}}}
\subsection{Method \code{createShinyServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$createShinyServer(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
