\name{MODIS.COVER.TYPE}
\alias{MODIS.COVER.TYPE}
\docType{data}
\title{ textual descriptions of land covers for Modis
 
}
\description{The modis land cover file consists of a series
  of values that represent land cover. Those values
  correspond to textual descriptions, like "water bodies"
  or "urban" or "evergreen"
 
}
\usage{ MODIS.COVER.TYPE }
\format{
  The format is:
 chr [1:18] "Water Bodies" "Evergreen Needleleaf Forest" ...
}
 
\examples{
print(MODIS.COVER.TYPE)
 
}
\keyword{constants}
