\name{NAVO.BZ.FILE}
\alias{NAVO.BZ.FILE}
\docType{data}
\title{ filename of the compressed version of coast data
 
}
\description{ The *.Bz2 filename of coast data
 
}
\usage{ NAVO.BZ.FILE }
\format{
  The format is:
 chr "lsmask-world8-var.dist5.5.nc.bz2"
}
 
\examples{
print(NAVO.BZ.FILE)
 
}
 
\keyword{filenames}
