\name{NAVO.FILE}
\alias{NAVO.FILE}
\docType{data}
\title{ File name of coast data
 
}
\description{ File name of the GHRSST distance to coast data
 
}
\usage{ NAVO.FILE }
\format{
  The format is:
 chr "lsmask-world8-var.dist5.5.nc"
}
 
\source{
 \url{https://www.ghrsst.org/data/ghrsst-data-tools/navo-ghrsst-pp-land-sea-mask/}
}
 
\examples{
print(NAVO.FILE)
 
}
 
\keyword{filenames}
