\name{NAVO_DIR}
\alias{NAVO_DIR}
\docType{data}
\title{ default directory for coast data
 
}
\description{ directory for coast data. The files related to
  GHRSST distnce from coast are placed in this directory
 
}
\usage{ NAVO_DIR }
\format{
  The format is:
 chr "NAVO"
}
 
\examples{
print(NAVO_DIR)
 
}
 
\keyword{directories}
