% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_transformer.R
\name{null_transformer}
\alias{null_transformer}
\title{Function to deal with nulls inside of glue}
\usage{
null_transformer(str = "null")
}
\arguments{
\item{str}{The string to treat as NULL}
}
\value{
The new string
}
\description{
Function to deal with nulls inside of glue
}
\keyword{internal}
