\name{NVG.FILE}
\alias{NVG.FILE}
\docType{data}
\title{ Filename for land use data. Sparse or no vegetation
 
}
\description{Filename for land use data. Sparse or no vegetation
 
}
\usage{ NVG.FILE }
\format{
  The format is:
 chr "NVG_2000.asc"
}
 
\examples{
print(NVG.FILE)
 
}
 
\keyword{filenames}
