% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderedList.R
\name{OrderedList}
\alias{OrderedList}
\title{Shiny interface for an ordered list of elements}
\description{
A modular shiny interface for managing an ordered list of elements
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrderedList-new}{\code{OrderedList$new()}}
\item \href{#method-OrderedList-copyData}{\code{OrderedList$copyData()}}
\item \href{#method-OrderedList-updateShinyUI}{\code{OrderedList$updateShinyUI()}}
\item \href{#method-OrderedList-createShinyUI}{\code{OrderedList$createShinyUI()}}
\item \href{#method-OrderedList-createShinyServer}{\code{OrderedList$createShinyServer()}}
\item \href{#method-OrderedList-addElement}{\code{OrderedList$addElement()}}
\item \href{#method-OrderedList-removeElement}{\code{OrderedList$removeElement()}}
\item \href{#method-OrderedList-moveElementUp}{\code{OrderedList$moveElementUp()}}
\item \href{#method-OrderedList-moveElementDown}{\code{OrderedList$moveElementDown()}}
\item \href{#method-OrderedList-clone}{\code{OrderedList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$new(name, buttonAddLabel = "Add")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-copyData"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-copyData}{}}}
\subsection{Method \code{copyData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$copyData(copy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-updateShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-updateShinyUI}{}}}
\subsection{Method \code{updateShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$updateShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-createShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-createShinyUI}{}}}
\subsection{Method \code{createShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$createShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-createShinyServer"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-createShinyServer}{}}}
\subsection{Method \code{createShinyServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$createShinyServer(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-addElement"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-addElement}{}}}
\subsection{Method \code{addElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$addElement(copy = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-removeElement"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-removeElement}{}}}
\subsection{Method \code{removeElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$removeElement(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-moveElementUp"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-moveElementUp}{}}}
\subsection{Method \code{moveElementUp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$moveElementUp(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-moveElementDown"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-moveElementDown}{}}}
\subsection{Method \code{moveElementDown()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$moveElementDown(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
