% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderedListPeople.R
\name{OrderedListPeople}
\alias{OrderedListPeople}
\title{An ordered list of people}
\description{
An ordered list of people
}
\section{Super class}{
\code{\link[metadata:OrderedList]{metadata::OrderedList}} -> \code{OrderedListPeople}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OrderedListPeople$new()}}
\item \href{#method-createElement}{\code{OrderedListPeople$createElement()}}
\item \href{#method-clone}{\code{OrderedListPeople$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="addElement">}\href{../../metadata/html/OrderedList.html#method-addElement}{\code{metadata::OrderedList$addElement()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="copyData">}\href{../../metadata/html/OrderedList.html#method-copyData}{\code{metadata::OrderedList$copyData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="createElementServer">}\href{../../metadata/html/OrderedList.html#method-createElementServer}{\code{metadata::OrderedList$createElementServer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="createElementUI">}\href{../../metadata/html/OrderedList.html#method-createElementUI}{\code{metadata::OrderedList$createElementUI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="createShinyServer">}\href{../../metadata/html/OrderedList.html#method-createShinyServer}{\code{metadata::OrderedList$createShinyServer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="createShinyUI">}\href{../../metadata/html/OrderedList.html#method-createShinyUI}{\code{metadata::OrderedList$createShinyUI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="moveElementDown">}\href{../../metadata/html/OrderedList.html#method-moveElementDown}{\code{metadata::OrderedList$moveElementDown()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="moveElementUp">}\href{../../metadata/html/OrderedList.html#method-moveElementUp}{\code{metadata::OrderedList$moveElementUp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="removeElement">}\href{../../metadata/html/OrderedList.html#method-removeElement}{\code{metadata::OrderedList$removeElement()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="removeElementUI">}\href{../../metadata/html/OrderedList.html#method-removeElementUI}{\code{metadata::OrderedList$removeElementUI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="updateShinyUI">}\href{../../metadata/html/OrderedList.html#method-updateShinyUI}{\code{metadata::OrderedList$updateShinyUI()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedListPeople$new(
  name,
  buttonAddLabel = "Add person",
  templatePerson = NULL,
  role = NULL,
  funding = FALSE,
  copy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createElement"></a>}}
\if{latex}{\out{\hypertarget{method-createElement}{}}}
\subsection{Method \code{createElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedListPeople$createElement(name, copy = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedListPeople$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
