\name{RAINFED.FILE}
\alias{RAINFED.FILE}
\docType{data}
\title{ filename of the cultivated land fed by rainfall
 
}
\description{filename of the cultivated land fed by rainfall
 
}
\usage{ RAINFED.FILE }
\format{
  The format is:
 chr "CULTRF_2000.asc"
}
 
\examples{
print(RAINFED.FILE)
 
}
 
\keyword{filenames}
