% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-to-eml.R
\name{read_metadata}
\alias{read_metadata}
\title{Connect to Google Drive and read SPI-Birds metadata sheet}
\usage{
read_metadata(email)
}
\arguments{
\item{email}{Character indicating the email address with which you have access to the SPI-Birds metadata sheet}
}
\description{
SPI-Birds metadata submissions are collected through Jotform and stored on Google Sheets.
}
\details{
Only used within \code{convert_to_eml()}.
}
