% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-to-eml.R
\name{set_study_site_ids}
\alias{set_study_site_ids}
\title{Set the siteID and studyID for a metadata entry}
\usage{
set_study_site_ids(entry)
}
\arguments{
\item{entry}{a single-row data frame representing a single metadata entry}
}
\description{
This function asks the user for a three-letter siteID, and creates an associated studyID.
If the metadata entry involves a new study at an existing site, a new studyID is automatically assigned.
If the metadata entry involves an update of an existing study, the user is asked to provide the associated studyID.
}
\details{
Only used within \link{convert_to_eml}.
}
