% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{site_codes}
\alias{site_codes}
\title{SPI-Birds site table: study site names and locations}
\format{
A data frame of 142 rows and 7 variables
\describe{
\item{siteID}{Character indicating the unique ID for a SPI-Birds study site.}
\item{siteName}{Character indicating the name of a SPI-Birds study site.}
\item{country}{Character indicating the name of the country where the study site is located.}
\item{countryCode}{Standard code for the country, using \href{https://www.iso.org/iso-3166-country-codes.html}{ISO 3166-1 alpha-2}.}
\item{decimalLatitude}{Geographic latitude of the geographic center of the study site in decimal degrees (WGS84).}
\item{decimalLongitude}{Geographic longitude of the geographic center of the study site in decimal degrees (WGS84).}
\item{locationAccordingTo}{Provider of the coordinate information. Either "metadataProvider" or "data".}
}
}
\usage{
site_codes
}
\description{
Geographic information and SPI-Birds-internal codes for all study sites.
}
\keyword{datasets}
