% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{SPI-Birds species table: species IDs, codes, and taxonomic ranks}
\format{
A data frame
\describe{
\item{speciesCode}{SPI-Birds' internal unique and persistent identifier for a species.}
\item{speciesID}{SPI-Birds' 6-letter species identifier. First three letters indicate the genus, last three letters indicate the species indicator.}
\item{speciesEURINGCode}{Species code used by \href{https://euring.org/}{EURING}. NA for species not included in EURING.}
\item{speciesCOLID}{Species ID used by the \href{https://www.catalogueoflife.org/}{Catalogue of Life}.}
\item{speciesEOLpageID}{Species page ID used by the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{speciesTSN}{Species taxonomic serial number (TSN) used by the \href{https://itis.gov}{Integrated Taxonomic Information System}.}
\item{kingdom}{Scientific name of the kingdom in which the species is identified according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{phylum}{Scientific name of the phylum in which the species is identified according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{class}{Scientific name of the class in which the species is identified according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{order}{Scientific name of the order in which the species is identified according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{family}{Scientific name of the family in which the species is identified according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{genus}{Scientific name of the genus in which the species is identified according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{specificEpithet}{Scientific name of the species epithet according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{scientificName}{Scientific name (genus + specific epithet) of the species according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{scientificNameAuthorship}{Authorship information of the scientific name, including date information if known. according to the \href{https://eol.org/}{Encyclopedia of Life}.}
\item{vernacularName}{English common name of species according to \href{https://https://www.wikidata.org/}{Wikidata}.}
}
}
\usage{
species_codes
}
\description{
Species information, including various taxonomic ranks, SPI-Birds-internal and external codes.
}
\keyword{datasets}
