% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{study_codes}
\alias{study_codes}
\title{SPI-Birds study table: studies and data custodians}
\format{
A data frame of 139 rows and 9 variables
\descrbe{
\item{studyID}{Character indicating the unique ID for a SPI-Birds study.}
\item{studyUUID}{Character indicating the univerisally unique identifier for the SPI-Birds study, to be referenced in the EML metadata file.}
\item{studyName}{Character indicating the name of the SPI-Birds study.}
\item{siteID}{Character indicating the unique ID for the SPI-Birds study site where the study is conducted.}
\item{custodianID}{Character indicating the unique ID for the person or organisation that has custody of the data coming from a study.}
\item{custodianName}{Character indicating the name of the person or organisation that has custody of the data coming from a study.}
\item{pipelineID}{Character indicating the unique ID for the pipeline that converts the data coming from a study into the SPI-Birds standard format.}
\item{data}{Logical indicating whether the data collected during the study are available through SPI-Birds.}
\item{standardFormat}{Logical indicating whether the data, if available through SPI-Birds, are standardised according to the SPI-Birds standard protocol.}
}
}
\usage{
study_codes
}
\description{
SPI-Birds-internal codes for all studies, the related study sites, and the status of the data and pipeline.
}
\keyword{datasets}
