% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplatePerson.R
\name{TemplatePerson}
\alias{TemplatePerson}
\title{A template for a person element}
\description{
A template for a person element
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TemplatePerson-new}{\code{TemplatePerson$new()}}
\item \href{#method-TemplatePerson-getPeopleIDs}{\code{TemplatePerson$getPeopleIDs()}}
\item \href{#method-TemplatePerson-getPersonInfo}{\code{TemplatePerson$getPersonInfo()}}
\item \href{#method-TemplatePerson-clone}{\code{TemplatePerson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePerson-new"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePerson-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePerson$new(people)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePerson-getPeopleIDs"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePerson-getPeopleIDs}{}}}
\subsection{Method \code{getPeopleIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePerson$getPeopleIDs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePerson-getPersonInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePerson-getPersonInfo}{}}}
\subsection{Method \code{getPersonInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePerson$getPersonInfo(personID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePerson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePerson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePerson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
