\name{URBAN.LAND.FILE}
\alias{URBAN.LAND.FILE}
\docType{data}
\title{ Filename for urban land use
 
}
\description{ filename for the urban land use data
 
}
\usage{ URBAN.LAND.FILE }
\format{
  The format is:
 chr "URB_2000.asc"
}
 
\source{
 url{http://www.iiasa.ac.at/Research/LUC/External-World-soil-database/
  HTML/LandUseShares.html?sb=9}
}
\references{Fischer, G., F. Nachtergaele, S. Prieler, H.T. van Velthuizen, 
   L. Verelst, D. Wiberg, 2008. Global Agro-ecological Zones Assessment 
   for Agriculture (GAEZ 2008). IIASA, Laxenburg, Austria and FAO, Rome, Italy.
  
}
\examples{
print(URBAN.LAND.FILE)
 
}
 
\keyword{filenames}
