% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_rmd_tags.R
\name{validate_rmd_tags}
\alias{validate_rmd_tags}
\title{Function to validate the tags in the Rmd document with the tags in RStudio Connect.}
\usage{
validate_rmd_tags(raw_rmd_tags, rsc_tags)
}
\arguments{
\item{raw_rmd_tags}{The output from \code{get_rmd_tags}; a hierarchical list with the tags in the Rmd.}

\item{rsc_tags}{The output from \code{get_all_rsc_tags}; an unordered list with the tags in RSC.}
}
\value{
a list containing two more lists; one with tags available in RSC and one with tags unavailable in RSC.
}
\description{
Function to validate the tags in the Rmd document with the tags in RStudio Connect.
}
\keyword{internal}
