% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\alias{add_experiments}
\alias{add_sites}
\alias{add_treatments}
\alias{add_species}
\alias{add_cultivars}
\alias{add_citations}
\title{Upload new metadata to BETYdb}
\usage{
add_experiments(user_id, dbcon, commit, experiments)

add_sites(user_id, dbcon, commit, sites)

add_treatments(user_id, dbcon, commit, treatments)

add_species(dbcon, cultivars)

add_cultivars(dbcon, commit, new_cultivars)

add_citations(user_id, dbcon, commit, citations)
}
\arguments{
\item{user_id}{User id for BETYdb account (integer)}

\item{dbcon}{Database connection, as created by [RPostgres::dbConnect]}

\item{commit}{Value indicating whether to commit database transactions (boolean, TRUE/FALSE)}

\item{experiments}{Experiments google sheet object}

\item{sites}{Sites google sheet object}

\item{treatments}{Treatments google sheet object}

\item{cultivars}{Cultivars google sheet object}

\item{new_cultivars}{New cultivars with specie ids to add to BETYdb (tibble)}

\item{citations}{Citations google sheet object}
}
\description{
Add new experiments, sites, treatments, cultivars, and citations to BETYdb. A user id
will need to be passed in when adding data to all tables but species and cultivars. Users can
choose to commit database transactions by setting the commit parameter to TRUE.
}
\examples{
\dontrun{

library(RPostgres)
library(googlesheets)
library(DBI)

# connect to test database
dbcon <- dbConnect(RPostgres::Postgres(),
host = 'localhost',
user = 'bety',
port = 5433)

# url to google sheet
url <- 'https://docs.google.com/spreadsheets/d/
1BPkZPfynZI9DB9bRJfa335uC_uNjST5jHVf-XUgSVvM/edit#gid=347205052'

# get key from URL
key <- extract_key_from_url(url)

# register sheet
gs_obj <- key \%>\% gs_key(lookup = FALSE)

# users sheet
users <- gs_obj \%>\% gs_read(ws = 'users')

# experiments sheet
experiments <- gs_obj \%>\% gs_read(ws = 'experiments')

# get user id from user login
bety_users <- tbl(dbcon, 'users') \%>\% collect()

user_id <- bety_users \%>\%
filter(login == users$login) \%>\%
select(id)

# add new experiments to BETYdb, commit database transaction
add_experiments(user_id$id, dbcon, commit = TRUE, experiments)

}

}
