% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-to-eml.R
\name{create_nested_taxonomy}
\alias{create_nested_taxonomy}
\title{Create EML's hierarchical taxonomy and refer to relevant taxon ids where available}
\usage{
create_nested_taxonomy(taxa)
}
\description{
Per studied species, this function will create a nested taxonomy including kingdom, phylum, class, order, family, genus, species
}
\details{
Only used within \code{convert_to_eml()}.
}
