% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-party.R
\name{create_party}
\alias{create_party}
\title{Create an additional party}
\usage{
create_party()
}
\value{
a structured list that can be directly plugged into an EML.xml as a creator, metadata provider, or contact.
}
\description{
The responsible party (creator), metadata provider, and contact in Jotform are limited to a single entry. In some instances, data contributors wish to add more than one party to these entities. This interactive function will help with the creation of that party.
}
