% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataNarrative.R
\name{MetadataNarrative}
\alias{MetadataNarrative}
\title{Shiny interface for managing narrative metadata for a data product}
\description{
A modular shiny interface for managing narrative metadata for a
  data product
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetadataNarrative$new()}}
\item \href{#method-copyData}{\code{MetadataNarrative$copyData()}}
\item \href{#method-createShinyUI}{\code{MetadataNarrative$createShinyUI()}}
\item \href{#method-createShinyServer}{\code{MetadataNarrative$createShinyServer()}}
\item \href{#method-updateShinyUI}{\code{MetadataNarrative$updateShinyUI()}}
\item \href{#method-clone}{\code{MetadataNarrative$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$new(name = "narrative", copy = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copyData"></a>}}
\if{latex}{\out{\hypertarget{method-copyData}{}}}
\subsection{Method \code{copyData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$copyData(metadataNarrative)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-createShinyUI}{}}}
\subsection{Method \code{createShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$createShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createShinyServer"></a>}}
\if{latex}{\out{\hypertarget{method-createShinyServer}{}}}
\subsection{Method \code{createShinyServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$createShinyServer(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-updateShinyUI}{}}}
\subsection{Method \code{updateShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$updateShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
