% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataNarrative.R
\name{MetadataNarrative}
\alias{MetadataNarrative}
\title{Shiny interface for managing narrative metadata for a data product}
\description{
A modular shiny interface for managing narrative metadata for a
  data product
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetadataNarrative-new}{\code{MetadataNarrative$new()}}
\item \href{#method-MetadataNarrative-read}{\code{MetadataNarrative$read()}}
\item \href{#method-MetadataNarrative-write}{\code{MetadataNarrative$write()}}
\item \href{#method-MetadataNarrative-createShinyUI}{\code{MetadataNarrative$createShinyUI()}}
\item \href{#method-MetadataNarrative-createShinyServer}{\code{MetadataNarrative$createShinyServer()}}
\item \href{#method-MetadataNarrative-updateShinyUI}{\code{MetadataNarrative$updateShinyUI()}}
\item \href{#method-MetadataNarrative-clone}{\code{MetadataNarrative$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataNarrative-new"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataNarrative-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$new(
  name = "narrative",
  path = "./01_input/meta",
  fileNames = c(title = "narr_title.txt", abstract = "narr_abstract.md", purpose =
    "narr_purpose.md", rights = "narr_rights.md", methods = "narr_methods.md")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataNarrative-read"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataNarrative-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$read(path = self$path, fileNames = self$fileNames)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataNarrative-write"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataNarrative-write}{}}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$write(path = self$path, fileNames = self$fileNames)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataNarrative-createShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataNarrative-createShinyUI}{}}}
\subsection{Method \code{createShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$createShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataNarrative-createShinyServer"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataNarrative-createShinyServer}{}}}
\subsection{Method \code{createShinyServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$createShinyServer(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataNarrative-updateShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataNarrative-updateShinyUI}{}}}
\subsection{Method \code{updateShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$updateShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataNarrative-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataNarrative-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataNarrative$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
