% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataPeople.R
\name{MetadataPeople}
\alias{MetadataPeople}
\title{Shiny interface for managing personnel on a data product}
\description{
A modular shiny interface for managing personnel on a data product
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetadataPeople-new}{\code{MetadataPeople$new()}}
\item \href{#method-MetadataPeople-read}{\code{MetadataPeople$read()}}
\item \href{#method-MetadataPeople-write}{\code{MetadataPeople$write()}}
\item \href{#method-MetadataPeople-updateShinyUI}{\code{MetadataPeople$updateShinyUI()}}
\item \href{#method-MetadataPeople-createShinyUI}{\code{MetadataPeople$createShinyUI()}}
\item \href{#method-MetadataPeople-createShinyServer}{\code{MetadataPeople$createShinyServer()}}
\item \href{#method-MetadataPeople-clone}{\code{MetadataPeople$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPeople-new"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPeople-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPeople$new(
  name = "people",
  templatePerson = NULL,
  templateFunding = NULL,
  path = "./01_input/meta",
  fileNames = c(creators = "ppl_creators.csv", contact = "ppl_contact.csv", others =
    "ppl_others.csv", funding = "ppl_funding.csv")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPeople-read"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPeople-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPeople$read(path = self$path, fileNames = self$fileNames)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPeople-write"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPeople-write}{}}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPeople$write(path = self$path, fileNames = self$fileNames)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPeople-updateShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPeople-updateShinyUI}{}}}
\subsection{Method \code{updateShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPeople$updateShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPeople-createShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPeople-createShinyUI}{}}}
\subsection{Method \code{createShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPeople$createShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPeople-createShinyServer"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPeople-createShinyServer}{}}}
\subsection{Method \code{createShinyServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPeople$createShinyServer(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPeople-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPeople-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPeople$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
