% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataPerson.R
\name{MetadataPerson}
\alias{MetadataPerson}
\title{Shiny interface for managing metadata for a person}
\description{
A modular shiny interface for managing metdata for a person
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetadataPerson-new}{\code{MetadataPerson$new()}}
\item \href{#method-MetadataPerson-copyData}{\code{MetadataPerson$copyData()}}
\item \href{#method-MetadataPerson-createShinyUI}{\code{MetadataPerson$createShinyUI()}}
\item \href{#method-MetadataPerson-updateShinyUI}{\code{MetadataPerson$updateShinyUI()}}
\item \href{#method-MetadataPerson-createShinyServer}{\code{MetadataPerson$createShinyServer()}}
\item \href{#method-MetadataPerson-clone}{\code{MetadataPerson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPerson-new"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPerson-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$new(
  name,
  templatePerson = NULL,
  role = NULL,
  funding = FALSE,
  copy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPerson-copyData"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPerson-copyData}{}}}
\subsection{Method \code{copyData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$copyData(metadataPerson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPerson-createShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPerson-createShinyUI}{}}}
\subsection{Method \code{createShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$createShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPerson-updateShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPerson-updateShinyUI}{}}}
\subsection{Method \code{updateShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$updateShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPerson-createShinyServer"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPerson-createShinyServer}{}}}
\subsection{Method \code{createShinyServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$createShinyServer(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetadataPerson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MetadataPerson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataPerson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
