% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderedList.R
\name{OrderedList}
\alias{OrderedList}
\title{Shiny interface for an ordered list of elements}
\description{
A modular shiny interface for managing an ordered list of elements
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OrderedList$new()}}
\item \href{#method-copyData}{\code{OrderedList$copyData()}}
\item \href{#method-updateShinyUI}{\code{OrderedList$updateShinyUI()}}
\item \href{#method-createShinyUI}{\code{OrderedList$createShinyUI()}}
\item \href{#method-createShinyServer}{\code{OrderedList$createShinyServer()}}
\item \href{#method-addElement}{\code{OrderedList$addElement()}}
\item \href{#method-createElementUI}{\code{OrderedList$createElementUI()}}
\item \href{#method-createElementServer}{\code{OrderedList$createElementServer()}}
\item \href{#method-removeElement}{\code{OrderedList$removeElement()}}
\item \href{#method-removeElementUI}{\code{OrderedList$removeElementUI()}}
\item \href{#method-moveElementUp}{\code{OrderedList$moveElementUp()}}
\item \href{#method-moveElementDown}{\code{OrderedList$moveElementDown()}}
\item \href{#method-clone}{\code{OrderedList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$new(name, buttonAddLabel = "Add")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copyData"></a>}}
\if{latex}{\out{\hypertarget{method-copyData}{}}}
\subsection{Method \code{copyData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$copyData(copy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-updateShinyUI}{}}}
\subsection{Method \code{updateShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$updateShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createShinyUI"></a>}}
\if{latex}{\out{\hypertarget{method-createShinyUI}{}}}
\subsection{Method \code{createShinyUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$createShinyUI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createShinyServer"></a>}}
\if{latex}{\out{\hypertarget{method-createShinyServer}{}}}
\subsection{Method \code{createShinyServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$createShinyServer(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addElement"></a>}}
\if{latex}{\out{\hypertarget{method-addElement}{}}}
\subsection{Method \code{addElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$addElement(copy = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createElementUI"></a>}}
\if{latex}{\out{\hypertarget{method-createElementUI}{}}}
\subsection{Method \code{createElementUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$createElementUI(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createElementServer"></a>}}
\if{latex}{\out{\hypertarget{method-createElementServer}{}}}
\subsection{Method \code{createElementServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$createElementServer(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeElement"></a>}}
\if{latex}{\out{\hypertarget{method-removeElement}{}}}
\subsection{Method \code{removeElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$removeElement(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeElementUI"></a>}}
\if{latex}{\out{\hypertarget{method-removeElementUI}{}}}
\subsection{Method \code{removeElementUI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$removeElementUI(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-moveElementUp"></a>}}
\if{latex}{\out{\hypertarget{method-moveElementUp}{}}}
\subsection{Method \code{moveElementUp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$moveElementUp(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-moveElementDown"></a>}}
\if{latex}{\out{\hypertarget{method-moveElementDown}{}}}
\subsection{Method \code{moveElementDown()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$moveElementDown(element)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
