% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderedListPeople.R
\name{OrderedListPeople}
\alias{OrderedListPeople}
\title{An ordered list of people}
\description{
An ordered list of people
}
\section{Super class}{
\code{\link[metadata:OrderedList]{metadata::OrderedList}} -> \code{OrderedListPeople}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrderedListPeople-new}{\code{OrderedListPeople$new()}}
\item \href{#method-OrderedListPeople-createElement}{\code{OrderedListPeople$createElement()}}
\item \href{#method-OrderedListPeople-clone}{\code{OrderedListPeople$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="addElement"><a href='../../metadata/html/OrderedList.html#method-OrderedList-addElement'><code>metadata::OrderedList$addElement()</code></a></li>
<li><span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="copyData"><a href='../../metadata/html/OrderedList.html#method-OrderedList-copyData'><code>metadata::OrderedList$copyData()</code></a></li>
<li><span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="createShinyServer"><a href='../../metadata/html/OrderedList.html#method-OrderedList-createShinyServer'><code>metadata::OrderedList$createShinyServer()</code></a></li>
<li><span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="createShinyUI"><a href='../../metadata/html/OrderedList.html#method-OrderedList-createShinyUI'><code>metadata::OrderedList$createShinyUI()</code></a></li>
<li><span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="moveElementDown"><a href='../../metadata/html/OrderedList.html#method-OrderedList-moveElementDown'><code>metadata::OrderedList$moveElementDown()</code></a></li>
<li><span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="moveElementUp"><a href='../../metadata/html/OrderedList.html#method-OrderedList-moveElementUp'><code>metadata::OrderedList$moveElementUp()</code></a></li>
<li><span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="removeElement"><a href='../../metadata/html/OrderedList.html#method-OrderedList-removeElement'><code>metadata::OrderedList$removeElement()</code></a></li>
<li><span class="pkg-link" data-pkg="metadata" data-topic="OrderedList" data-id="updateShinyUI"><a href='../../metadata/html/OrderedList.html#method-OrderedList-updateShinyUI'><code>metadata::OrderedList$updateShinyUI()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedListPeople-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedListPeople-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedListPeople$new(
  name,
  buttonAddLabel = "Add person",
  templatePerson = NULL,
  role = NULL,
  funding = FALSE,
  copy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedListPeople-createElement"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedListPeople-createElement}{}}}
\subsection{Method \code{createElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedListPeople$createElement(name, copy = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedListPeople-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedListPeople-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedListPeople$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
